tutorials/requirements.txt
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.0.4
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
biopython==1.73
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
checkbox-support==0.39.0
command-not-found==0.3
cryptography==2.3.1
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.2
defer==1.0.6
defusedxml==0.5.0
distro==1.3.0
distro-info===0.18ubuntu0.18.04.1
Django==2.1.2
docutils==0.14
entrypoints==0.2.3
evdev==1.1.2
feedparser==5.2.1
guacamole==0.9.2
Guake==3.4.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.7
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.3
jeepney==0.4
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==15.1.0
keyrings.alt==3.1
Kivy==1.11.0.dev0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.5
macaroonbakery==1.2.1
Mako==1.0.7
MarkupSafe==1.0
mate-hud==18.4.8
mate-tweak==18.4.16
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
mysqlclient==1.3.13
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.4
nose==1.3.7
notebook==5.7.6
numpy==1.16.2
oauth==1.0.1
oauthlib==2.1.0
olefile==0.46
onboard==1.4.1
padme==1.1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.0.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.3.0
pipenv==2018.10.13
plainbox==0.38.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.4.2
ptyprocess==0.6.0
pyasn1==0.4.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.6.4
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.2.2
pyRFC3339==1.1
pyrsistent==0.14.11
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.33
python-xapp==1.0.0
python-xlib==0.23
pytz==2018.5
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.4.3
reportlab==3.5.9
requests==2.19.1
requests-unixsocket==0.1.5
roman==2.0.0
scipy==1.2.1
screen-resolution-extra==0.0.0
SecretStorage==3.1.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
stevedore==1.29.0
sympy==1.3
system-service==0.3
systemd-python==234
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24
usb-creator==0.3.3
virtualenv==16.0.0
virtualenv-clone==0.4.0
wadllib==1.3.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xdiagnose==3.8.8
xkit==0.0.0
XlsxWriter==0.9.6
youtube-dl==2018.3.14
zope.interface==4.5.0
